requirements.txt
elasticsearch==8.10.1
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
webvtt-py==0.4.6
cachetools==5.3.1